ALLOCATING: Note that a single identifier is sufficient for a single software
Since the introduction of the 'class' key, many authors have requested
up to 4 identifiers for a single software, because the software implemented
a tracker, an igate, a digipeater and a weather station.  Make it clear
that a single device identifier is sufficient for a single software,
even if it has many features.